+++ Coimbra/TimerHandle.cs
@@ -7,7 +7,6 @@ namespace Coimbra
     /// Handle for a timer from <see cref="ITimerService"/>.
     /// </summary>
     [Preserve]
-    [Serializable]
     public readonly struct TimerHandle : IEquatable<TimerHandle>
     {
         public readonly Guid Guid;