+++ Coimbra/ValueChangeHandler.cs
@@ -1,9 +1,0 @@
-﻿using JetBrains.Annotations;
-
-namespace Coimbra
-{
-    /// <summary>
-    /// Generic delegate for value changes.
-    /// </summary>
-    public delegate void ValueChangeHandler<in T>([CanBeNull] T oldValue, [CanBeNull] T newValue);
-}