+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 - Added EventRef and APIs on IEvenService to use it.
 - Added call to Dispose inside OnDestroy in MonoBehaviourServiceBase.
+- Added GameObjectBehaviour class that works similar to extending GameObject class itself.
 - Changed MonoBehaviourServiceBase to allow multiple on the same GameObject.
 - Changed EventListenerHandler to use the new EventRef type.
+- Changed GameObjectEventListenerComponent functionality to the new GameObjectBehaviour class.
 
 ## [4.0.0] - 2022-03-25